refactor(targeting): move group member store into targeting domain
