+++ src/inbound-handler.ts
@@ -13,7 +13,7 @@ import {
   buildLearningContextBlock,
   isLearningEnabled,
 } from "./feedback-learning-service";
-import { formatGroupMembers, noteGroupMember } from "./group-members-store";
+import { formatGroupMembers, noteGroupMember } from "./targeting/group-members-store";
 import { setCurrentLogger } from "./logger-context";
 import { prepareMediaInput, resolveOutboundMediaType } from "./media-utils";
 import {
