+++ tests/unit/group-members-store.test.ts
@@ -2,7 +2,7 @@ import * as fs from "node:fs";
 import * as os from "node:os";
 import * as path from "node:path";
 import { afterEach, describe, expect, it } from "vitest";
-import { formatGroupMembers, noteGroupMember } from "../../src/group-members-store";
+import { formatGroupMembers, noteGroupMember } from "../../src/targeting/group-members-store";
 import { resolveNamespacePath } from "../../src/persistence-store";
 
 function makeStorePath(): { rootDir: string; storePath: string } {
